+++ src/app/modules/society/flat-list/flat-list.component.ts
@@ -1,18 +1,22 @@
-import { Component, OnDestroy, OnInit } from '@angular/core';
+import { Component, OnDestroy, OnInit, TemplateRef, ViewChild } from '@angular/core';
 import { FormControl, FormGroup, Validators } from '@angular/forms';
-import { ActivatedRoute, Router } from '@angular/router';
+import { ActivatedRoute } from '@angular/router';
 import { SocietyService } from '../../../services/society.service';
 import { Observable, Subject, take, takeUntil } from 'rxjs';
 import { IBuilding, IFlat, ISociety, IUIControlConfig, IUIDropdownOption } from '../../../interfaces';
 import { FlatTypeList, FlatTypes, PERMISSIONS } from '../../../constants';
 import { LoginService } from '../../../services/login.service';
+import { MatDialog, MatDialogRef } from '@angular/material/dialog';
+import { DialogService } from '../../../services/dialog.service';
+import { WindowService } from '../../../services/window.service';
+import { ListBase } from '../../../directives/list-base.directive';
 
 @Component({
   selector: 'app-flat-list',
   templateUrl: './flat-list.component.html',
   styleUrl: './flat-list.component.scss'
 })
-export class FlatListComponent implements OnInit, OnDestroy {
+export class FlatListComponent extends ListBase implements OnInit, OnDestroy {
 
   societyId?: string;
   buildingId?: string;
@@ -20,6 +24,9 @@ export class FlatListComponent implements OnInit, OnDestroy {
   buildings: IBuilding[] = [];
   building?: IBuilding;
   flats: IFlat[] = [];
+
+  @ViewChild('flatTemplate') flatTemplate!: TemplateRef<any>;
+  currentDialogRef: MatDialogRef<any> | null = null;
 
   fb = new FormGroup({
     _id: new FormControl<string | undefined>(''),
@@ -36,7 +43,8 @@ export class FlatListComponent implements OnInit, OnDestroy {
       flatNumber: new FormControl<string | null>(null, [Validators.required]),
       floor: new FormControl<number>(1, [Validators.required]),
       flatType: new FormControl<FlatTypes>(FlatTypes['1BHK'], [Validators.required])
-    })
+    }),
+    appendFloorNumber: new FormControl<boolean[]>([])
   });
   floorControl = new FormControl<string | null>(null);
 
@@ -152,6 +160,10 @@ export class FlatListComponent implements OnInit, OnDestroy {
       required: 'Flat Number is required',
     }
   };
+  appendFloorNumberConfig: IUIControlConfig = {
+    id: 'appendFloorNumber',
+    label: 'Append Floor Number',
+  };
 
   selectedTab = 'addFlat';
   tabsOptions: IUIDropdownOption[] = [
@@ -163,6 +175,9 @@ export class FlatListComponent implements OnInit, OnDestroy {
       value: 'autoGen',
       label: 'Autogenerate Flats'
     },
+  ];
+  appendFloorNumberOptions: IUIDropdownOption[] = [
+    { label: 'Append Floor Number with Flat Number ?', value: true }
   ];
   defaultFilter: IUIDropdownOption = {
     label: 'All',
@@ -196,9 +211,9 @@ export class FlatListComponent implements OnInit, OnDestroy {
   }
 
   get pageTitle(): string | undefined {
-    if (!this.society) return 'Flat Manager';
+    if (!this.society) return 'Flats';
 
-    return this.society.societyName + ' Flats Manager'
+    return 'Flats: ' + this.society.societyName
   }
 
   get flatTypeOptions(): IUIDropdownOption<FlatTypes>[] {
@@ -229,10 +244,12 @@ export class FlatListComponent implements OnInit, OnDestroy {
 
   constructor(
     private route: ActivatedRoute,
-    private router: Router,
     private societyService: SocietyService,
-    private loginService: LoginService
-  ) { }
+    private loginService: LoginService,
+    private dialog: MatDialog,
+    dialogService: DialogService,
+    private windowService: WindowService
+  ) { super(dialogService) }
 
   ngOnInit(): void {
     this.societyId = this.route.snapshot.paramMap.get('id')!;
@@ -340,11 +357,11 @@ export class FlatListComponent implements OnInit, OnDestroy {
         this.loadFlats(societyId, building._id);
       });
 
-      this.floorControl.valueChanges
+    this.floorControl.valueChanges
       .pipe(takeUntil(this.isComponentActive))
       .subscribe({
         next: response => {
-          
+
         }
       });
   }
@@ -362,14 +379,48 @@ export class FlatListComponent implements OnInit, OnDestroy {
     }
   }
 
+  resetFlatForm() {
+    this.fb.get('addFlats')?.setValue({ flatNumber: null, floor: 1, flatType: FlatTypes['1BHK'] }, { emitEvent: false });
+    this.fb.get('autogenerateForm')?.setValue({ fromFloor: 1, toFloor: 1, startFlat: 1, endFlat: 1, flatType: FlatTypes['1BHK'] }, { emitEvent: false })
+    this.fb.get('appendFloorNumber')?.setValue([]);
+    this.selectedTab = 'addFlat';
+    this.handleTabChange(this.selectedTab);
+  }
+
+  getDialogWidth(): string {
+    let width = '50%';
+    switch (this.windowService.mode.value) {
+      case 'mobile': width = '90%'; break;
+      case 'tablet': width = '70%'; break;
+      case 'desktop': width = '60%'; break
+    }
+    return width;
+  }
+  openAddDialog() {
+    this.resetFlatForm();
+    this.currentDialogRef = this.dialog.open(this.flatTemplate, {
+      width: this.getDialogWidth(),
+      panelClass: 'building-form-dialog'
+    });
+    this.currentDialogRef.afterClosed().subscribe(() => {
+      this.currentDialogRef = null;
+      this.resetFlatForm();
+    });
+  }
+
+  closeDialog() {
+    this.currentDialogRef?.close();
+  }
+
   addFlat() {
     if (this.fb.invalid) return;
 
     const formValue = this.fb.getRawValue();
+    const connectFloorToFlatNumber = formValue.appendFloorNumber && formValue.appendFloorNumber.length > 0 && formValue.appendFloorNumber[0] === true;
     let obs: Observable<any> | undefined;
     if (this.selectedTab === 'addFlat') {
       const payload = {
-        flatNumber: formValue.addFlats?.flatNumber,
+        flatNumber: (connectFloorToFlatNumber ? formValue.addFlats.floor ?? '' : '') + (formValue.addFlats?.flatNumber ?? ''),
         buildingId: formValue.building,
         societyId: formValue.society?._id ?? '',
         flatType: formValue.addFlats?.flatType,
@@ -392,11 +443,14 @@ export class FlatListComponent implements OnInit, OnDestroy {
       const endFlat = (autogenerateForm?.endFlat ?? 0) * 1;
       const arrFloors = Array.from({ length: toFloor - fromFloor + 1 }, (_, i) => fromFloor + i);
       const arrFlatNumbers = Array.from({ length: endFlat - startFlat + 1 }, (_, i) => startFlat + i);
+      const digitLength = endFlat > 9 ? endFlat.toString().length : 2;
 
       const flats = arrFloors.reduce((arrFlats, floorNumber) => {
         arrFlatNumbers.forEach(flatNumber => {
+          const paddedFlat = flatNumber.toString().padStart(digitLength, '0');
+
           arrFlats.push({
-            flatNumber: flatNumber.toString(),
+            flatNumber: (connectFloorToFlatNumber ? floorNumber : '') + paddedFlat,
             flatType,
             floor: floorNumber,
             societyId: formValue.society?._id ?? '',
@@ -425,6 +479,7 @@ export class FlatListComponent implements OnInit, OnDestroy {
           this.loadFlats(societyId, buildingId);
           this.fb.get('autogenerateForm')?.reset();
           this.fb.get('addFlats')?.reset();
+          this.closeDialog();
         }
       });
   }
@@ -437,8 +492,10 @@ export class FlatListComponent implements OnInit, OnDestroy {
       .join(', ')
   }
 
-  deleteFlat(flat: IFlat) {
+  async deleteFlat(flat: IFlat) {
     if (!this.societyId) return;
+
+    if (!await this.dialogService.confirmDelete('Delete Flat', `Are you sure you want to delete flat ${flat.flatNumber} ?`)) return;
 
     this.societyService.deleteFlat(this.societyId, flat._id)
       .pipe(take(1))
@@ -448,6 +505,18 @@ export class FlatListComponent implements OnInit, OnDestroy {
           this.loadFlats(this.societyId ?? '', buildingId ?? undefined);
         },
       })
+  }
+
+  deleteOneRecord(id: string) {
+    if (!this.societyId) return;
+
+    return this.societyService.deleteFlat(this.societyId, id);
+  }
+
+  refreshList() {
+    if (!this.societyId || !this.selectedBuilding) return;
+
+    this.loadFlats(this.societyId, this.selectedBuilding);
   }
 
   ngOnDestroy(): void {
